requirements.txt
streamlit
faiss-cpu
joblib
pandas
numpy
thefuzz
requests
huggingface_hub

